requirements.txt
boto3==1.34.162
boto3==1.35.37
langchain==0.3.7
langchain_aws==0.2.6
numpy==2.1.3
openai==1.53.0
opensearch_py==2.5.0
peft==0.12.0
python-dotenv==1.0.1
Requests==2.32.3
requests_aws4auth==1.1.1
sagemaker==2.232.3
sentence_transformers==3.2.1
streamlit==1.39.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
